requirements.txt
django == 3.1.7
psycopg2 == 2.8.6
djangorestframework == 3.12.2